Verificar escuela de data relacionada a ciclos académicos que se muestra y registra. También solucionar bug en validador al momento de validad longitudes singulares, siendo esta validación errónea se procedía a intentar validar un rango, lo cual no debía de hacerlo, ya no lo hace. También solucionar la edición de perfíl.

+++ alembic/versions/cf6bd44b1c51_crear_tabla_horario_profesor.py
@@ -19,7 +19,7 @@ depends_on = None
 def upgrade():
     op.create_table('horario_profesor',
                     sa.Column('id', sa.Integer, primary_key=True, autoincrement=True),
-                    sa.Column('ciclo_id', sa.SmallInteger, nullable=False),
+                    sa.Column('ciclo_id', sa.Integer, nullable=False),
                     sa.Column('profesor_id', sa.BigInteger, nullable=False),
                     sa.Column('dia_clase', sa.SmallInteger, nullable=False),
                     sa.Column('hora_comienzo', sa.SmallInteger, default=0, nullable=False),
